Add a virtual site element
The symbol is VS (should it be EP?). The mass and atomic number are both zero.
It evaluates to boolean False (and is the only element to do so), which supports
*some* aspect of the old API (since None also evaluates to boolean False).

This also simplifies some of the logic (getting rid of special-casing virtual
sites where it was done) and changes the Topology creators to use "virtual"
instead of "None" for unrecognized elements.