Add more unit tests for hoomdxml reader
There are several different scenarios possible when reading a hoomdxml
file (no chains, only chains, mixture of chains and "ions").
A bug was only revealed when reading a file with no bonds, i.e., all
ions.
Further, the bonds were being stored erroneously, as the networkx
package returns the edges (bonds) as an unordered set.
This commit adds tests to avoid these problems in the future.