Add an accelerated version of the distance/displacement triclinic calculation.
It uses the vectorized version to get *a* close value, then loops over all 27
adjacent unit cells without explicit SSE instructions to get the nearest
neighbor. This slow part is only executed for triclinic cells that are NOT
orthorhombic, so it only slows down calculation for the cases that didn't work
correctly before.

Also disable optimized calculation of angles and dihedrals when a
non-orthorhombic cell is present and periodic is True, since the optimized
kernel only exists for the orthorhombic case. Print out a warning about what is
done (i.e., opt is set to False to ensure correctness over speed+uncertainty).

Still need to add some tests.