[xtc|trr] added xdrlib2 (seek support).
This method will read an integer per frame to determine the byte offsets between
frames. It will only be called once the user wants to know the lengths of the
trajectory or he/she demands random access (seeking). The offsets are accessible
via a property of the XTC|TRRTrajectoryFile and could in principle cached.