[xtc|trr] added seek support to provide fast random access.
This method will read an integer per frame to determine the byte offsets between
frames for XTC files. TRR files on the other hand store the sizes of a frame in
a private data structure in xdrlib (which has been exposed to handle this).

For small XTC files (<= 9 atoms), the coordinates are written uncompressed,
which is reflected in the offset calculation.

Note that the amount of bytes needed for frames can be figured out by reading
the original source code of xdrlib (GROMACS).

The byte offset array will only be constructed if requested by the user,if he/she
wants to know the lengths of the trajectory or he/she demands random access
(seeking). The offsets are accessible via a property of the
XTC|TRRTrajectoryFile and cached (stored to disk and restored via the setter).