Added entire periodic table of vdW radii. For atoms where no value is known, we use 0.2 nm. For many group 1 and group 2 metals and chlorine, their ionic radii were used.
The shrake_rupley function in sasa.py was updated to accept an
optional dictionary of values to replace the built-in radii, if the
user so desires.

Tests were updated to compare with gromacs. The tests on alanine
dipeptide were difficult to get to match and I had to further reduce
the number of significant digits. I think this is because gromacs
uses a slightly different algorithm to place the points and
because of the precision of the trajectory and comparison files.