Fix bug in load with atom_indices and frame args
Both bugs occur for topology formats that can also be trajectories
because MDTraj attempts to apply the args twice. When passing `frame`,
it's simply the `TypeError` reported in #1226. When it's `atom_indices`,
the topology gets sliced once when loading the topology and then again
when loading the trajectory. This does not cause any problems when
slicing from 0:n but fails when slicing indices past that.

Resolves #1226