Fix tranform calculation and Transform object to be more general.
Currenlty, if you create a transform object with compute_tranformation,
you will get a Transform object that will only be able to correctly
transform groups of atoms that are the same as the target, i.e. you
can't align an entire system including water when you just compute
the tranformation matrix on protein.