Add calc_directors or order.py
The director for each molecule is an intermediate step in the
calculation of the nematic order parameter. Often times, this quantity
is useful, e.g., calculating the tilt angle of a monolayer. This commit
refactors the order code, so that compute_directors() is a user-facing
function. This is done by calling compute_directors() directly in
compute_nematic_order(). The "parsing" of the indices is also refactored
out into a separate function.