Add compute_directors to and tests for order.py
The director for each molecule is an intermediate step in the
calculation of the nematic order parameter. Often times, this quantity
is useful, e.g., calculating the tilt angle of a monolayer. This commit
refactors the order code, so that compute_directors() is a user-facing
function. This is done by calling compute_directors() directly in
compute_nematic_order(). The "parsing" of the indices is also refactored
out into a separate function.

This commit also adds unit tests for compute_nematic_order() and
compute_directors(). An example monolayer was added to
testing/reference; this monolayer had chains aligned exactly parallel to
the z-axis.