[formats/xtc] use byte offsets between frames for stride if avail
This should avoid a lot of IO for high stride arguments,
because unwanted frames are just seeked over.
This feature relies on the byte offsets, which have to be provided
or explictly computed by accessing the offsets property once.
If no bytes offset is available, the current behaviour is untouched.

Added tests.