[formats/mol2] add more checks to element parsing
In DataFrame.atype.apply we get a list for Python2, but a
str for Python3 - so we need to distinguish this.

Moreover we need to check for orbital hybridization ($E.sp_x)
and 'weird' element names, which are not specified in SYBL nor GAFF,
e.g. CT, OS, CK etc.
The latter is ensured via trying to parse the element first with its
two letter code (Li etc.), if that fails, fall back to a one letter
code. This succeeds at least for the provided test examples.

Fixes #1387