Minor addition: create standard bonds
Not sure if this will work as intended, but I added one line to automatically create the standard bonds list. This is necessary for other commands such as the hydrogen bonding wernet_nilsson() or baker_hubbard() functions. Without this change, an empty list would be returned upon using those functions.