ValueError Addition : Check topology.bonds > 0 (#1513)
Check that topology.bonds has a length greater than zero in the "_get_bond_triplets" function; otherwise, raise an error. Having a topology.bonds list is necessary to run the  hydrogen bond commands: baker_hubbard and wernet_nilsson; however, no explicit error message is given and an empty list is returned after running the aforementioned functions. This added check isn't robust as it only checks to see if there is at least one bond in topology.bonds, ... but it gets the job done!