Adds hydrogens to is_sidechain and is_backbone
Previously backbone hydrogens (conventionally named 'HA' for the alpha carbon hydrogen and 'H' for the nitrogen hydrogen) were not considered in the atom.is_sidechain and atom.is_backbone methods. This could lead to errant behavior (even within internal mdtraj functions); for example, previously compute_contacts using the 'sidechain' scheme would consider ALL of the hydrogens (including backbone hydrogens) to be part of the 'sidechain' for the purposes of calculating minimum distances.