[core.Topology.from_dataframe] BUGFIX: force new residue when creating new chain
Previously, a new residue was detected on the condition of new resname
or new resSeq. This leaves out the case where subsequent residues
have same resname and same resSeq but different chainID.

In the code, the previous_chainID has been already overwritten when
the new residue detection takes place, so we simply force a new residue
on a new chain, which we detect with c.n_atoms=0.